requirements.txt
# python3 version == 3.8.10 64bit

# !* == not vital for this project but might be for org project
# org == package project origin

Flask==1.1.4
gunicorn==19.10.0
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
pytz==2019.3
requests==2.25.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
Flask-JSGlue==0.3.1 
python-dotenv==0.18.0 # PYTHON VIRTUAL ENV
PyJWT==2.1.0 # JWT AUTH
marshmallow-sqlalchemy==0.19.0
marshmallow==2.21.0
SQLAlchemy-Utils==0.37.9 # !* ( org == license_check_api >> flask_skeleton )
flask-marshmallow==0.14.0
Flask-Mail==0.9.1